Changed the iterator i to be of type size_t instead of int

+++ src/io_service_pool.cpp
@@ -3,7 +3,7 @@
 io_service_pool::io_service_pool(const size_t num_threads) : w(new work(*this))
 {
 	reserve(num_threads);
-	for (int i = 0; i < num_threads; ++i)
+	for (size_t i = 0; i < num_threads; ++i)
 	{
 		emplace_back(async(launch::async, [&]()
 		{
